+++ seasonthon_team_25_fe/lib/ui/bank/ui/financial_product_list.dart
@@ -71,7 +71,7 @@ class _FinancialProductListPageState
               child: Align(
                 alignment: Alignment.topLeft,
                 child: CoinBalanceChip(
-                  balance: balance,
+                  balance: balance.toInt(),
                   backgroundColor: AppColors.sk_25,
                   textColor: AppColors.primarySky,
                 ),
