+++ seasonthon_team_25_fe/lib/ui/bank/ui/financial_product_sign_up.dart
@@ -95,7 +95,7 @@ class _FinancialProductSignUpPageState
               Align(
                 alignment: Alignment.topLeft,
                 child: CoinBalanceChip(
-                  balance: balance,
+                  balance: balance.toInt(),
                   backgroundColor: AppColors.sk_25,
                   textColor: AppColors.primarySky,
                 ),
