+++ seasonthon_team_25_fe/lib/ui/bank/ui/terminated_savings_page.dart
@@ -48,7 +48,7 @@ class _TerminatedSavingsPageState extends ConsumerState<TerminatedSavingsPage> {
             Align(
               alignment: Alignment.topLeft,
               child: CoinBalanceChip(
-                balance: balance,
+                balance: balance.toInt(),
                 backgroundColor: AppColors.sk_25,
                 textColor: AppColors.primarySky,
               ),
